requirements.txt
fastapi
pydantic
python-dotenv
# python-multipart
SQLAlchemy
psycopg2
